Address cubic round 2 + add DeepL adapter with capabilities metadata
Cubic: connect dialog is non-dismissible while validating (cancel could
orphan an in-flight connect that still persists the key); audit changeset
translates the user_id FK violation instead of raising during concurrent
account deletion; audit actor_id widened to bigint to match users.id.

DeepL joins as the seventh provider (owner-decided): /v2/usage validation,
DeepL-Auth-Key auth, :fx-suffixed keys routed to the free host. Provider
metadata gains the immutable capabilities list for all seven adapters per
the platform plan's lane policy.

+++ priv/repo/migrations/20260720143500_create_ai_integration_audits.exs
@@ -8,7 +8,7 @@ defmodule Storyarn.Repo.Migrations.CreateAiIntegrationAudits do
       add :user_id, references(:users, on_delete: :nilify_all)
       # Immutable snapshot of the acting user's id. No FK on purpose — it must
       # survive account deletion (pseudonymous, not direct PII).
-      add :actor_id, :integer, null: false
+      add :actor_id, :bigint, null: false
       add :provider, :string, null: false
       add :action, :string, null: false
       add :metadata, :map, default: %{}, null: false
